Refactor ProjectionMode
