+++ src/utils/load-video.tsx
@@ -1,5 +1,6 @@
 /** @jsx createElementEntity */
-import { createElementEntity, ProjectionMode } from "../utils/jsx-entity";
+import { createElementEntity } from "../utils/jsx-entity";
+import { ProjectionMode } from "./projection-mode";
 import { VideoTexture } from "three";
 import { renderAsEntity } from "../utils/jsx-entity";
 import { loadVideoTexture } from "../utils/load-video-texture";
