+++ src/utils/projection-mode.ts
@@ -1,0 +1,4 @@
+export enum ProjectionMode {
+  FLAT = "flat",
+  SPHERE_EQUIRECTANGULAR = "360-equirectangular"
+}
